+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Check origin of verification method when verifying activity signature.
 - Write message to log when server stops.
+- Copy `role.permissions` to `role.permissions_names` in `Account` Mastodon API entity.
 
 ## [3.20.0] - 2025-03-15